+++ tests/testthat/test-reduce.R
@@ -113,21 +113,11 @@ test_that("accumulate() uses vctrs to simplify results", {
 test_that("accumulate() does not fail when input can't be simplified", {
   expect_identical(accumulate(list(1L, 2:3), ~ .y), list(1L, 2:3))
   expect_identical(accumulate(list(1, "a"), ~ .y), list(1, "a"))
-  expect_identical(accumulate(1:3, ~ .y), 1:3)
-  expect_identical(accumulate(list(identity), ~ .y), list(identity))
-  expect_identical(accumulate(mtcars, ~ .y), as.list(mtcars))
 })
 
-test_that("accumulate() does not simplify data frame rowwise", {
-  out <- accumulate(
-    1L,
-    ~ data.frame(new = .y),
-    .init = data.frame(new = 0L)
-  )
-  exp <- list(data.frame(new = 0L), data.frame(new = 1L))
-  expect_identical(out, exp)
+test_that("accumulate() does fail when simpification is required", {
+  expect_snapshot(accumulate(list(1, "a"), ~ .y, .simplify = TRUE), error = TRUE)
 })
-
 
 # reduce2 -----------------------------------------------------------------
 
@@ -162,8 +152,8 @@ test_that("basic accumulate2() works", {
   paste2 <- function(x, y, sep) paste(x, y, sep = sep)
 
   x <- c("a", "b", "c")
-  expect_equal(accumulate2(x, c("-", "."), paste2), list("a", "a-b", "a-b.c"))
-  expect_equal(accumulate2(x, c(".", "-", "."), paste2, .init = "x"), list("x", "x.a", "x.a-b", "x.a-b.c"))
+  expect_equal(accumulate2(x, c("-", "."), paste2), c("a", "a-b", "a-b.c"))
+  expect_equal(accumulate2(x, c(".", "-", "."), paste2, .init = "x"), c("x", "x.a", "x.a-b", "x.a-b.c"))
 })
 
 test_that("can terminate accumulate2() early", {
@@ -177,8 +167,8 @@ test_that("can terminate accumulate2() early", {
   }
 
   x <- c("a", "b", "c")
-  expect_equal(accumulate2(x, c("-", "."), paste2), list("a", "a-b"))
-  expect_equal(accumulate2(x, c(".", "-", "."), paste2, .init = "x"), list("x", "x.a", "x.a-b"))
+  expect_equal(accumulate2(x, c("-", "."), paste2), c("a", "a-b"))
+  expect_equal(accumulate2(x, c(".", "-", "."), paste2, .init = "x"), c("x", "x.a", "x.a-b"))
 })
 
 test_that("accumulate2() forces arguments (#643)", {